feat: move to more common storybook
