+++ packages/containers/.storybook/main.js
@@ -1,0 +1,4 @@
+module.exports = {
+	stories: ['../src/**/*.stories.@(js)'],
+	addons: ['@storybook/addon-links', '@storybook/addon-essentials', '@storybook/addon-toolbars'],
+};
