+++ packages/containers/.storybook/settings.json
@@ -1,0 +1,209 @@
+{
+  "props": {
+    "appheaderbar": {
+      "app": "Hello Test"
+    },
+    "Translate(HeaderBar)#default": {
+      "logo": { "name": "appheaderbar:logo", "isFull": true },
+      "brand": { "label": "DATA STREAMS" },
+      "notification": { "name": "appheaderbar:notification" }
+    },
+    "ActionButton#first": {
+      "label": "First",
+      "icon": "talend-streams",
+      "payload": {
+        "type": "MENU_"
+      }
+    },
+    "ActionButton#second": {
+      "label": "Second",
+      "icon": "talend-dataprep",
+      "payload": {
+        "type": "MENU_"
+      }
+    },
+    "ActionButton#configuration": {
+      "label": "Configuration",
+      "icon": "talend-cog",
+      "payload": {
+        "type": "MENU_"
+      }
+    }
+  },
+  "actions": {
+    "show:about": {
+      "label": "Show",
+      "payload": {
+        "type": "ABOUT_DIALOG_SHOW",
+        "url": "https://tdp.us.cloud.talend.com/api/version"
+      }
+    },
+    "appheaderbar:logo": {
+      "icon": "talend-logo"
+    },
+    "appheaderbar:notification": {
+      "icon": "talend-world"
+    },
+    "menu:first": {
+      "label": "First",
+      "icon": "talend-streams",
+      "payload": {
+        "type": "MENU_"
+      }
+    },
+    "menu:second": {
+      "label": "Second",
+      "icon": "talend-dataprep",
+      "payload": {
+        "type": "MENU_"
+      }
+    },
+    "menu:third": {
+      "label": "Configuration",
+      "icon": "talend-cog",
+      "payload": {
+        "type": "MENU_"
+      }
+    },
+    "menu:fourth": {
+      "label": "Upload",
+      "icon": "talend-upload",
+      "displayMode": "file",
+      "payload": {
+        "type": "UPLOAD"
+      }
+    },
+    "object:add": {
+      "label": "Add",
+      "icon": "talend-plus-circle",
+      "bsStyle": "primary",
+      "payload": {
+        "type": "APP_OBJECT_ADD"
+      }
+    },
+    "object:remove": {
+      "label": "Remove",
+      "icon": "talend-trash",
+      "bsStyle": "danger",
+      "payload": {
+        "type": "APP_OBJECT_REMOVE"
+      }
+    },
+    "object:upload": {
+      "label": "Upload",
+      "icon": "talend-upload",
+      "displayMode": "file",
+      "payload": {
+        "type": "UPLOAD"
+      }
+    },
+    "object:edit": {
+      "label": "Edit",
+      "icon": "talend-pencil",
+      "payload": {
+        "type": "APP_OBJECT_EDIT"
+      }
+    },
+    "object:delete": {
+      "label": "Delete",
+      "icon": "talend-trash",
+      "payload": {
+        "type": "APP_OBJECT_DELETE"
+      }
+    },
+    "object:confirm:dialog": {
+      "label": "Remove",
+      "bsStyle": "primary",
+      "id": "object:confirm:dialog",
+      "actionCreator": "confirm:dialog"
+    },
+    "object:hide:dialog": {
+      "label": "Cancel",
+      "id": "object:hide:dialog",
+      "actionCreator": "cancel:hide:dialog"
+    },
+    "menu:items": {
+      "id": "menu:items",
+      "displayMode": "dropdown",
+      "label": "my items",
+      "itemsExpression": "getItems",
+      "actionCreator": "object:view"
+    },
+    "menu:items-id": {
+      "id": "menu:items-id",
+      "displayMode": "dropdown",
+      "label": "my items",
+      "actionIds": ["menu:first", "menu:second"],
+      "actionCreator": "object:view"
+    },
+    "menu:href": {
+      "id": "menu:href",
+      "label": "Talend",
+      "target": "_blank",
+      "href": "//www.talend.com"
+    },
+    "menu:dropdown-href": {
+      "id": "menu:dropdown-href",
+      "displayMode": "dropdown",
+      "label": "my items",
+      "actionIds": ["menu:href"],
+      "actionCreator": "object:view"
+    },
+    "dialog:delete:validate": {
+      "id": "dialog:delete:validate",
+      "label": "Yes",
+      "bsStyle": "danger",
+      "actionCreator": "confirm:dialog"
+    },
+    "dialog:delete:cancel": {
+      "id": "dialog:delete:cancel",
+      "label": "No",
+      "className": "btn-inverse",
+      "actionCreator": "cancel:hide:dialog"
+    },
+    "action:overlay:component": {
+      "id": "action:overlay:component",
+      "label": "overlay with component",
+      "overlayComponent": "ComponentOverlay",
+      "overlayComponentProps": {
+        "customProps": "customProps"
+      },
+      "overlayPlacement": "bottom",
+      "payload": { "type": "BUTTON_OVERLAY" }
+    },
+    "action:icon:toggle": {
+      "icon": "talend-panel-opener-right",
+      "id": "action:icon:toggle",
+      "label": "Click me to toggle",
+      "tooltipPlacement": "top",
+      "activeExpression": { "id": "isFlagExpression", "args": ["action:icon:creator:flag"] },
+      "payload": { "type": "TOGGLE_FLAG_TYPE", "flagId": "action:icon:creator:flag" }
+    },
+    "show:guidedTour": {
+      "label": "Start guided tour",
+      "payload": {
+        "type": "GUIDED_TOUR_SHOW"
+      }
+    },
+    "subheaderbar:action-sharing": {
+      "id": "subheaderbar:action-sharing",
+      "label": "sharing",
+      "renderType": "action",
+      "actionCreator": "subheaderbar:display-sharing",
+      "icon": "talend-share-alt",
+      "bsStyle": "link",
+      "hideLabel": true,
+      "overlay": false
+    },
+    "subheaderbar:action-bubbles": {
+      "id": "subheaderbar:action-bubbles",
+      "label": "bubbles",
+      "renderType": "action",
+      "actionCreator": "subheaderbar:display-bubbles",
+      "icon": "talend-bubbles",
+      "bsStyle": "link",
+      "hideLabel": true,
+      "overlay": false
+    }
+  }
+}
