+++ packages/containers/.storybook/webpack.config.js
@@ -5,4 +5,20 @@
 // IMPORTANT
 // When you add this file, we won't add the default configurations which is similar
 // to "React Create App". This only has babel loader to load JavaScript.
-module.exports = require('../../../.storybook/webpack.config');
+const CopyPlugin = require('copy-webpack-plugin');
+const getConfig = require('../../../.storybook/webpack.config');
+
+module.exports = (...args) => {
+	const config = getConfig(...args);
+	config.plugins.push(
+		new CopyPlugin({
+			patterns: [
+				{
+					from: './.storybook/settings.json',
+					to: '/',
+				},
+			],
+		}),
+	);
+	return config;
+};
